refactor(contracts): split into abstractions and implementations
