--- src/abstractWrappers/index.js
@@ -1,4 +1,0 @@
-import Arbitrator from './arbitrator'
-import ArbitrableContracts from './arbitrableContracts'
-
-export { Arbitrator, ArbitrableContracts }
