--- src/contractWrappers/PNK/index.js
@@ -1,3 +1,0 @@
-import _PinakionPOC from './PinakionPOC'
-
-export const PinakionPOC = _PinakionPOC
