+++ src/contracts/abstractions/index.js
@@ -1,0 +1,4 @@
+import Arbitrator from './Arbitrator'
+import ArbitrableContracts from './Arbitrable'
+
+export { Arbitrator, ArbitrableContracts }
