+++ src/contracts/implementations/PNK/index.js
@@ -1,0 +1,3 @@
+import PinakionPOC from './PinakionPOC'
+
+export { PinakionPOC }
