+++ src/contracts/implementations/RNG/index.js
